+++ examples/mobile/expo-54-go/providers/Web3Provider.tsx
@@ -1,4 +1,4 @@
-import { sophonOSTestnet } from '@sophon-labs/account-core';
+import { DataScopes, sophonOSTestnet } from '@sophon-labs/account-core';
 import {
   SophonContextProvider,
   SophonWagmiConnector,
@@ -17,8 +17,7 @@ export const Web3Provider = ({ children }: { children: React.ReactNode }) => {
       partnerId="123b216c-678e-4611-af9a-2d5b7b061258"
       insets={insets}
       locale="en"
-
-      // dataScopes={[DataScopes.email, DataScopes.apple]}
+      dataScopes={[DataScopes.email, DataScopes.apple]}
       // requestedCapabilities={[Capabilities.WALLET_CONNECT]}
     >
       <SophonWagmiProvider>