+++ packages/account-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sophon-labs/account-core
 
+## 3.0.0-next.38
+
+### Patch Changes
+
+- Updating react native signature to use owner instead of contract
+
 ## 3.0.0-next.37
 
 ### Patch Changes