+++ packages/account-provider/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sophon-labs/account-eip6963
 
+## 3.0.0-next.38
+
+### Patch Changes
+
+- Updating react native signature to use owner instead of contract
+- Updated dependencies
+  - @sophon-labs/account-communicator@3.0.0-next.38
+  - @sophon-labs/account-core@3.0.0-next.38
+
 ## 3.0.0-next.37
 
 ### Patch Changes